backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0

# Ollama integration
langchain-ollama>=0.1.0

# Vector database
chromadb>=0.4.0

# Database ORM
sqlalchemy>=2.0.0
